+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Old_code/TeleOp/RedTeleOp.java
@@ -1,9 +1,7 @@
-package org.firstinspires.ftc.teamcode.TeleOp;
+package org.firstinspires.ftc.teamcode.Old_code.TeleOp;
 
-import com.acmerobotics.dashboard.FtcDashboard;
 import com.acmerobotics.dashboard.config.Config;
 import com.qualcomm.hardware.bosch.BNO055IMU;
-import com.qualcomm.hardware.rev.RevTouchSensor;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
 import com.qualcomm.robotcore.hardware.DcMotor;
@@ -12,16 +10,14 @@ import com.qualcomm.robotcore.hardware.DcMotorSimple;
 import com.qualcomm.robotcore.hardware.DigitalChannel;
 import com.qualcomm.robotcore.hardware.Gamepad;
 import com.qualcomm.robotcore.hardware.Servo;
-import com.qualcomm.robotcore.hardware.TouchSensor;
 import com.qualcomm.robotcore.util.ElapsedTime;
 
-import org.firstinspires.ftc.robotcore.external.Telemetry;
 import org.firstinspires.ftc.robotcore.external.navigation.AngleUnit;
 import org.firstinspires.ftc.teamcode.classes.TapeMeasure;
 
 @Config
-@TeleOp(name = "BLUE Freight Frenzy TeleOp", group = "BLUE")
-public class BlueTeleOp extends LinearOpMode {
+@TeleOp(name = "RED Freight Frenzy TeleOp", group = "RED")
+public class RedTeleOp extends LinearOpMode {
     public static double gateClosed = .1, gateOpen = .37, bottomClosed = 1, bottomOpen = 0, quasiGateOpen = .2, straight, strafe, rotation;
     public static int BOTTOM = 175, MIDDLE = 250, HIGH = 350;
     public int targetPos = 0;
@@ -39,7 +35,7 @@ public class BlueTeleOp extends LinearOpMode {
         boolean toggle = false, armState = false, toggle2 = false, toggle3 = false;
         boolean rumble1 = true, rumble2 = true, rumble3 = true;
         String currentTarget = "top";
-        TapeMeasure tapeMeasure = new TapeMeasure(hardwareMap, gamepad1,gamepad2, TapeMeasure.Side.BLUE);
+        TapeMeasure tapeMeasure = new TapeMeasure(hardwareMap, gamepad1,gamepad2, TapeMeasure.Side.RED);
         Thread tapeMeasureThread = new Thread(tapeMeasure);
 
         ElapsedTime timer = new ElapsedTime();
@@ -105,9 +101,9 @@ public class BlueTeleOp extends LinearOpMode {
         timer.reset();
         while (opModeIsActive()) {
 
-            if (gamepad1.right_bumper) {
+            if (gamepad1.right_bumper || gamepad2.y) {
                 intake.setPower(1);
-            } else if (gamepad1.left_bumper) {
+            } else if (gamepad1.left_bumper || gamepad2.a) {
                 intake.setPower(-1);
             } else {
                 intake.setPower(0);
