+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/TeleDrive_LinearOpMode.java
@@ -379,15 +379,15 @@ public class TeleDrive_LinearOpMode extends LinearOpMode {
                     position2 = .43;
                 }
                 if(gamepad1.b){
-                   //For quick position setting and other tweaks.
+                   //For quick position setting and other cheats.
                 }
                 if(gamepad1.x){
-                    //For quick position setting and other tweaks.
+                    //For quick position setting and other cheats.
                 }
                 if(gamepad1.y){
-                    //For quick position setting and other tweaks.
+                    //For quick position setting and other cheats.
                 }
-
+                //showing values
                 servo.setPosition(position);
                 servo2.setPosition(position2);
                 telemetry.addData("Horizontal Position", "%5.2f", position);
